stamp: get config method to load Supabase env
